skill-d-create-issues/requirements.txt
requests==2.31.0
python-dateutil==2.9.0
pytz==2024.1
pyyaml==6.0.1
fastapi==0.110.0
uvicorn==0.29.0